refactor package structure and rename rank identifiers for consistency
