+++ surf-bitmap-provider-bitmaps/src/main/kotlin/dev/slne/surf/bitmap/bitmaps/Bitmaps.kt
@@ -1,7 +1,7 @@
 package dev.slne.surf.bitmap.bitmaps
 
 import dev.slne.surf.bitmap.api.BitmapProvider
-import dev.slne.surf.bitmap.bitmaps.bitmaps.ClanDefault
+import dev.slne.surf.bitmap.bitmaps.bitmaps.clans.ClanDefault
 import dev.slne.surf.bitmap.bitmaps.bitmaps.ranks.*
 
 enum class Bitmaps(
