+++ surf-bitmap-provider-bitmaps/src/main/kotlin/dev/slne/surf/bitmap/bitmaps/bitmaps/ranks/RanksOtherPlayer.kt
@@ -3,7 +3,7 @@ package dev.slne.surf.bitmap.bitmaps.bitmaps.ranks
 import dev.slne.surf.bitmap.api.BitmapProvider
 
 object RanksOtherPlayer : BitmapProvider(
-    name = "other_player",
+    name = "otherPlayer",
     a = '꒴',
     b = '꒵',
     c = '꒶',
